requirements.txt
asgiref==3.3.4
Django==3.2.2
django-cleanup==5.2.0
django-phonenumber-field==5.1.0
phonenumbers==8.12.22
Pillow==8.2.0
python-dotenv==0.17.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
